% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.align_celltypes}
\alias{.align_celltypes}
\title{Align a proportions/celltype matrix's columns to a target set of cell types}
\usage{
.align_celltypes(mat, celltypes, what = "proportions")
}
\description{
Align a proportions/celltype matrix's columns to a target set of cell types
}
\keyword{internal}
