% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.align_sites}
\alias{.align_sites}
\title{Align a reference/theta matrix's rows to a target set of site ids}
\usage{
.align_sites(mat, site_ids, what = "reference")
}
\description{
Align a reference/theta matrix's rows to a target set of site ids
}
\keyword{internal}
