% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure_bioc.R
\name{.ensure_bioc_installed}
\alias{.ensure_bioc_installed}
\title{Check that a set of Bioconductor packages are installed, stopping with
an actionable message listing everything missing if not.}
\usage{
.ensure_bioc_installed(pkgs)
}
\arguments{
\item{pkgs}{character vector of Bioconductor package names.}
}
\description{
Check that a set of Bioconductor packages are installed, stopping with
an actionable message listing everything missing if not.
}
\keyword{internal}
