% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_annotation.R
\name{.txdb_pkg_for_genome}
\alias{.txdb_pkg_for_genome}
\title{Which UCSC `TxDb` annotation package covers a given genome build.}
\usage{
.txdb_pkg_for_genome(genome)
}
\description{
Which UCSC `TxDb` annotation package covers a given genome build.
}
\keyword{internal}
